refactores LogIn component so login function is in separate module
